Normalize paths relative to basePath in all cases
Paths that include the basePath as set in options on the APISpec
instance were normalized to not have the basePath as part of each path.
However, this didn't work with the Flask plugin as normalization occured
before the plugin generated it's URLs.

This fixes the issue by also normalizing paths after plugin processing in
`add_path`.

Fixes #78.