feat: support fields metadata with non-string fields
This commit solves the issue that I wrongly filed to flask-apispec repo:
https://github.com/jmcarp/flask-apispec/issues/196
Using desert to convert a dataclass/attrs class to a Schema,
the metadata contains a non string key (_DesertSentinel)
that causes the metadata2properties method to raise an AttributeError.
This quick fix uses str(key) in the two dict comprehensions
to get the string representation of the key and return the dictionary.