do not add non docummented path
As a user I want to be able to have route without docstring and generate a
valid yaml file.

Path without any documentation are returned like this:
```
paths:
  /random_pet: {}
```

However, if the path contains a parameter we have:

```
paths:
  /get_by_id/{id}: {}
```
Which is not ok, according to swagger editor:

`Declared path parameter "id" needs to be defined as a path parameter at either
the path or operation level`

This patch ignore route without any operation defined, so route without
documentation are not included.